HasRequest and HasValidator traits
