+++ src/Illuminate/Validation/Concerns/HasValidator.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Illuminate\Auth\Concerns;
+
+trait HasValidator
+{
+    /**
+     * The current validation.
+     *
+     * @var \Illuminate\Contracts\Validation\Validator
+     */
+    protected $validator;
+
+    /**
+     * Set the current validator.
+     *
+     * @param  \Illuminate\Contracts\Validation\Validator  $validator
+     * @return $this
+     */
+    public function setValidator($validator)
+    {
+        $this->validator = $validator;
+
+        return $this;
+    }
+}
