+++ src/Illuminate/Validation/Rules/Can.php
@@ -5,9 +5,12 @@ namespace Illuminate\Validation\Rules;
 use Illuminate\Contracts\Validation\Rule;
 use Illuminate\Contracts\Validation\ValidatorAwareRule;
 use Illuminate\Support\Facades\Gate;
+use Illuminate\Validation\Concerns\HasValidator;
 
 class Can implements Rule, ValidatorAwareRule
 {
+    use HasValidator;
+
     /**
      * The ability to check.
      *
@@ -21,13 +24,6 @@ class Can implements Rule, ValidatorAwareRule
      * @var array
      */
     protected $arguments;
-
-    /**
-     * The current validator instance.
-     *
-     * @var \Illuminate\Validation\Validator
-     */
-    protected $validator;
 
     /**
      * Constructor.
@@ -69,18 +65,5 @@ class Can implements Rule, ValidatorAwareRule
         return $message === 'validation.can'
             ? ['The :attribute field contains an unauthorized value.']
             : $message;
-    }
-
-    /**
-     * Set the current validator.
-     *
-     * @param  \Illuminate\Validation\Validator  $validator
-     * @return $this
-     */
-    public function setValidator($validator)
-    {
-        $this->validator = $validator;
-
-        return $this;
     }
 }
